Fix bug and change store handler interface API
Fix bug causing the service store projection version
to not be updated to the latest version during a sync scan
when the last scanned event wasn't applied because
the service isn't subscribing to it.
Add a new method `UpdateProjectionVersion` to the
store handler interface which will be called automatically
to set the projection version in the store whenever necessary.
Also, remove the responsibility for updating the
projection version from the `ApplyEventXXX` methods
of the store handler interface to avoid
unnecessary code duplication.